% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exhep.R
\name{InitHEP}
\alias{InitHEP}
\title{Initialize HExpPridict module}
\usage{
InitHEP()
}
\value{
An R6 class object.
}
\description{
Initialize HExpPridict module analysis. It can generate an R6 class object
integrating all the analysis information.
}
\details{
ExpoHExpPredict module is designed  to predict blood concentrations of
chemicals and prioritize chemicals of health concern.
}
\examples{
res <- InitHEP()
}
\author{
Mingliang Fang, Bin Wang,(corresponding author)
}
