% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exHExpPred.R
\name{LoadHEP}
\alias{LoadHEP}
\title{Load data file for HExpPridict module}
\usage{
LoadHEP(PID, UseExample = "default", DataPath=NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitHEP.}

\item{UseExample}{chr. Method of uploading data. If "default",user should upload their own data files,
or use "example#1" provided by this module.}

\item{DataPath}{chr. Input file directory, e.g. "D:/test/eg_hep.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for HExpPridict module.
}
\examples{
res <- InitHEP()
   res = LoadHEP(PID = res$PID, UseExample = "example#1")
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
