% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exHExpPred.R
\name{PredBlood}
\alias{PredBlood}
\title{Predict blood concentrations of chemicals}
\usage{
PredBlood(PID, OutPath="default",MC = "F",N=100)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitHEP.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{MC}{lgl. T (or TRUE) and F (or FALSE). Whether to perform Mentocaro simulation.}

\item{N}{num. Times of Mentocaro simulation.}
}
\value{
A list object containing the prediction results.
}
\description{
Predict blood concentrations of chemicals
}
\examples{
res = InitHEP()
   res1 = LoadHEP(PID = res$PID,UseExample = "example#1")
   res2 = ConvToExpoID(PID = res$PID)
   res3 = PredBlood(PID = res$PID, OutPath="default", MC ='F',N=1000)
}
\author{
Mingliang Fang, Bin Wang (corresponding author)
}
