% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exHExpPred.R
\name{VizPredBlood}
\alias{VizPredBlood}
\title{Visualize the PredBlood}
\usage{
VizPredBlood(PID,OutPath="default",Layout="forest",
    Brightness="light",Palette= "science")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}

\item{OutPath}{chr. Output file directory. e.g. "D:/test". It should be noted that the slash symbol is "/", not "\".
If "default", the current working directory will be set.}

\item{Layout}{chr. Visualization layout. Available options include "forest" and "boxplot".}

\item{Brightness}{chr. Visualization brightness. Available options include "light" and "dark".}

\item{Palette}{chr. Visualization palette. Available options include "default1", "default2"
and several journal preference styles (i.e., cell, nature, science, lancet, nejm, and jama).}
}
\value{
A list object containing the plot of the biological link. This plot can be further processed using ggplot2 package.
}
\description{
Visualize the PredBlood.
}
\examples{
res = InitHEP()
   res1 = LoadHEP(PID = res$PID,UseExample = "example#1")
   res2 = ConvToExpoID(PID = res$PID)
   res3 = PredBlood(PID = res$PID, OutPath="default", MC ='F', N=1000)
   res4 = VizPredBlood(PID = res$PID, OutPath ="default",Layout="forest",
   Brightness="light",Palette= "science" )
}
\author{
Mingliang Fang, Ning Gao, Bin Wang (corresponding author)
}
